# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_AuthSession_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Code"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthSession"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          }
        },
        "required": ["code", "msg"],
        "title": "ApiResponse[AuthSession]",
        "type": "object"
      },
...
      "ApiResponse_NoneType_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Code"
          },
          "data": {
            "title": "Data",
            "type": "null"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          }
        },
        "required": ["code", "msg"],
        "title": "ApiResponse[NoneType]",
        "type": "object"
      },
...
      "AuthSession": {
        "properties": {
          "access_token_expire": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token Expire"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "refresh_token_expire": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token Expire"
          }
        },
        "title": "AuthSession",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginModel": {
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": ["username", "password"],
        "title": "LoginModel",
        "type": "object"
      },
...
      "RegisterModel": {
        "properties": {
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": ["name", "password"],
        "title": "RegisterModel",
        "type": "object"
      },
...
      "ResetPasswordModel": {
        "properties": {
          "new_password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": ["token", "new_password"],
        "title": "ResetPasswordModel",
        "type": "object"
      },
      "ResetPasswordRequest": {
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 254,
            "title": "Email",
            "type": "string"
          }
        },
        "required": ["email"],
        "title": "ResetPasswordRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CookieSession": {
        "description": "HttpOnly access-token cookie issued by Sun World authentication endpoints.",
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Sun World API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthSession_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": ["auth"]
      }
    },
...
    "/auth/refresh_token": {
      "post": {
        "operationId": "refresh_token_auth_refresh_token_post",
        "parameters": [
          {
            "in": "header",
            "name": "refresh-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Refresh-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthSession_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Token",
        "tags": ["auth"]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthSession_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": ["auth"]
      }
    },
    "/auth/reset_password": {
      "post": {
        "operationId": "reset_password_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Password",
        "tags": ["auth"]
      }
    },
    "/auth/reset_password/request": {
      "post": {
        "operationId": "request_reset_password_auth_reset_password_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Reset Password",
        "tags": ["auth"]
      }
    },
    "/auth/session": {
      "get": {
        "operationId": "session_status_auth_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthSession_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Session Status",
        "tags": ["auth"]
      }
    },
...
  }
}